add ctx to the pgx log
